consolidate duplicate tests
The tested code paths are after the deferred annotations handling so
can be moved to a more logical location.
