+++ tests/service/test_decorators.py
@@ -1,5 +1,7 @@
 from typing import Annotated
 
+import pytest
+
 from dbus_fast import PropertyAccess
 from dbus_fast import introspection as intr
 from dbus_fast.annotations import (
@@ -160,3 +162,36 @@ def test_interface_introspection():
     assert len(methods) == 1
     assert len(signals) == 1
     assert len(properties) == 2
+
+
+def test_bad_dbus_signature_annotation():
+    # Error message should tell the user how to fix it (use typing.Annotated with
+    # DBusSignature) and what they did wrong (used 'str').
+    with pytest.raises(
+        ValueError, match=r".*typing\.Annotated with DBusSignature.*<class 'str'>.*"
+    ):
+
+        class BadInterface(ServiceInterface):  # pyright: ignore[reportUnusedClass]
+            def __init__(self):
+                super().__init__("bad.interface")
+
+            @dbus_method()
+            def bad_method(self, what: str) -> str:
+                return what
+
+
+def test_bad_dbus_signature_annotation2():
+    # Error message should tell the user how to fix it (use DBusSignature) and
+    # what they did wrong (used typing.Annotated[...] without DBusSignature).
+    with pytest.raises(
+        ValueError,
+        match=r".*type must include a DBusSignature annotation.*typing.Annotated\[str, 's'\].*",
+    ):
+
+        class BadInterface(ServiceInterface):  # pyright: ignore[reportUnusedClass]
+            def __init__(self):
+                super().__init__("bad.interface")
+
+            @dbus_method()
+            def bad_method(self, what: Annotated[str, "s"]) -> Annotated[str, "s"]:
+                return what
