+++ tests/service/test_methods.py
@@ -257,35 +257,3 @@ async def test_methods(interface_class):
     await asyncio.wait_for(bus1.wait_for_disconnect(), timeout=1)
     await asyncio.wait_for(bus2.wait_for_disconnect(), timeout=1)
 
-
-def test_bad_dbus_signature_annotation():
-    # Error message should tell the user how to fix it (use typing.Annotated with
-    # DBusSignature) and what they did wrong (used 'str').
-    with pytest.raises(
-        ValueError, match=r".*typing\.Annotated with DBusSignature.*<class 'str'>.*"
-    ):
-
-        class BadInterface(ServiceInterface):  # pyright: ignore[reportUnusedClass]
-            def __init__(self):
-                super().__init__("bad.interface")
-
-            @dbus_method()
-            def bad_method(self, what: str) -> str:
-                return what
-
-
-def test_bad_dbus_signature_annotation2():
-    # Error message should tell the user how to fix it (use DBusSignature) and
-    # what they did wrong (used typing.Annotated[...] without DBusSignature).
-    with pytest.raises(
-        ValueError,
-        match=r".*type must include a DBusSignature annotation.*typing.Annotated\[str, 's'\].*",
-    ):
-
-        class BadInterface(ServiceInterface):  # pyright: ignore[reportUnusedClass]
-            def __init__(self):
-                super().__init__("bad.interface")
-
-            @dbus_method()
-            def bad_method(self, what: Annotated[str, "s"]) -> Annotated[str, "s"]:
-                return what
